Set up generic OptionsMenu class
