+++ CMakeLists.txt
@@ -217,6 +217,7 @@ set(Impacto_Src
         src/ui/sysmesbox.cpp
         src/ui/backlogmenu.cpp
         src/ui/tipsmenu.cpp
+        src/ui/optionsmenu.cpp
         src/ui/widgets/label.cpp
         src/ui/widgets/button.cpp
         src/ui/widgets/backlogentry.cpp
@@ -533,6 +534,7 @@ set(Impacto_Header
         src/ui/sysmesbox.h
         src/ui/backlogmenu.h
         src/ui/tipsmenu.h
+        src/ui/optionsmenu.h
         src/ui/widget.h
         src/ui/widgets/label.h
         src/ui/widgets/button.h
